requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.24.0=py36_0
blas=1.0=mkl
ca-certificates=2019.3.9=hecc5488_0
certifi=2019.3.9=py36_0
cffi=1.12.3=py36h2e261b9_0
chardet=3.0.4=py36_1
cryptography=2.3.1=py36hc365091_0
cudatoolkit=10.0.130=0
cudnn=7.3.1=cuda10.0_0
cymem=2.0.2=py36hfd86e86_0
cytoolz=0.9.0.1=py36h14c3975_1
dill=0.2.9=py36_0
django=2.1.7=py36_0
ftfy=5.5.1=py_0
idna=2.8=py36_0
intel-openmp=2019.3=199
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=8.2.0=hdf63c60_1
mkl=2018.0.3=1
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
msgpack-numpy=0.4.3.2=py36_0
msgpack-python=0.6.1=py36hfd86e86_1
murmurhash=1.0.2=py36he6710b0_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
nltk=3.4=py36_1
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
openssl=1.0.2r=h14c3975_0
pandas=0.24.2=py36he6710b0_0
pip=19.0.3=py36_0
plac=0.9.6=py36_0
preshed=2.0.1=py36he6710b0_0
pudb=2018.1=py36_1000
pycparser=2.19=py36_0
pygments=2.3.1=py_0
pyopenssl=19.0.0=py36_0
pysocks=1.6.8=py36_0
python=3.6.4=hc3d631a_3
python-dateutil=2.8.0=py36_0
pytorch-cpu=1.0.1=py3.6_cpu_2
pytz=2018.9=py36_0
readline=7.0=h7b6447c_5
regex=2019.03.09=py36h7b6447c_0
requests=2.21.0=py36_0
scikit-learn=0.20.1=py36h4989274_0
scipy=1.1.0=py36hfa4b5c9_1
setuptools=40.8.0=py36_0
six=1.12.0=py36_0
spacy=2.0.16=py36h962f231_0
sqlite=3.27.2=h7b6447c_0
thinc=6.12.1=py36h4989274_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
tqdm=4.31.1=py36_1
ujson=1.35=py36h14c3975_0
urllib3=1.24.1=py36_0
urwid=2.0.1=py36h14c3975_1000
wcwidth=0.1.7=py36_0
wheel=0.33.1=py36_0
wrapt=1.10.11=py36h14c3975_2
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
